section3/requirements.txt
category-encoders==2.3.0
certifi==2020.6.20
click==8.0.3
Flask==2.0.2
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe @ file:///opt/concourse/worker/volumes/live/c9141381-1dba-485b-7c96-99007bf7bcfd/volume/markupsafe_1621528150226/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.22.1
pandas==1.4.0
patsy==0.5.2
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
statsmodels==0.13.1
threadpoolctl==3.0.0
Werkzeug @ file:///tmp/build/80754af9/werkzeug_1635505089296/work
